Refactor ManifestComparison into separate file
Move some functions into methods on ManifestComparison
